DiskConnectHandle.ReadAt was not handling reads at the end of the disk properly.  If the blocks being read (#53)
extended beyond the end of the disk, an EOF would be returned with 0 bytes being read.  This would cause
uploads of the data to abort prematurely and lose data.  Modified DiskConnectHandle.ReadAt to check the actual
size of the disk/snapshot being read and adjust the size of the buffer so that the available blocks are read and
returned.

Signed-off-by: Dave Smith-Uchida <dsmithuchida@vmware.com>